+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [BUGFIX] Do not propagate attributes from Errors and LongTasks to Views.
 - [IMPROVEMENT] Upgrade to PLCrashReporter 1.11.1.
+- [FEATURE] Report session sample rate to the backend with RUM events. See [#1410][]
 
 # 2.0.0 / 31-07-2023
 
@@ -491,6 +492,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1331]: https://github.com/DataDog/dd-sdk-ios/pull/1331
 [#1328]: https://github.com/DataDog/dd-sdk-ios/pull/1328
 [#1355]: https://github.com/DataDog/dd-sdk-ios/pull/1355
+[#1410]: https://github.com/DataDog/dd-sdk-ios/pull/1410
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu